+++ NEWS.md
@@ -6,6 +6,7 @@
 - Updated the reporting template to work with the development version of `onbrand`
 - Changed terminal messaging to use `cli`
 - Added termination criteria in estimation output
+- Fixed wrapping issue in Fortran output where unwrappable strings resulted in infinite loop
 
 # ubiquity 2.0.0